Remove hostname validation
The hostname part of a remote archive is not necessarily a host
name: it can be a name of a Host section in .ssh/config file, or
the like.  Verifying it with getaddrinfo is not appropriate.  Besides,
if the hostname doesn't resolve or cannot be reached, it will
be diagnosed by rmt_open anyway.

* am/system.m4: Don't check for getaddrinfo.
* lib/rtapelib.c: Remove hostname validation.